Fix waiting for "on boot" in Android 4.4
Turns out the property we wait for isn't always set on all devices.
On Android 4.4 devices, we have to look for another property. And in
case even that is not set, we have timeout failback to not prevent the
rest of the system to continue.

This fixes the completion of lxc-android-config job, which in turns
fixes zram on some devices.